backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
openai==1.3.0
python-multipart==0.0.6
aiofiles==23.1.0
